requirements.txt
hydra-core
numpy
pandas
tqdm
torch
torchvision
tensorboard
